feat: Add functions for deleting and updating messages
This commit adds two new functions to the Teleprompter namespace: DeleteMessage and UpdateMessage. These functions allow for deleting and updating prompt messages respectively. Additionally, a new function SendUpdates is added to send batches of prompt delete and update messages asynchronously.

The changes aim to enhance the functionality of the Teleprompter HTTP SDK.